Rename: apis 폴더 분리 및 README 작성
