+++ .gitignore
@@ -17,8 +17,6 @@
 /.next/
 /out/
 
-.env
-.env.local
 
 # production
 /build
