+++ src/app/api/auth/callback/kakao/apis/index.ts
@@ -1,0 +1,4 @@
+import getSocialKakao from '@/app/api/auth/callback/kakao/apis/getSocialKakao';
+import useKakaoLogin from '@/app/api/auth/callback/kakao/apis/useKakaoLogin';
+
+export { getSocialKakao, useKakaoLogin };
