+++ src/app/api/auth/callback/line/apis/useLineLogin.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 
-import getSocialLine from '@/app/api/auth/callback/line/getSocialLine';
+import getSocialLine from '@/app/api/auth/callback/line/apis/getSocialLine';
 
 const useLineLogin = () => useMutation({ mutationFn: getSocialLine });
